refactor: refactor submission model
